+++ package.json
@@ -35,8 +35,6 @@
     "@angular/compiler-cli": "^6.1.0",
     "@angular/language-service": "^6.1.0",
     "@material/button": "^0.40.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
     "angular-router-loader": "^0.8.5",
     "angular2-template-loader": "^0.6.2",
@@ -46,15 +44,7 @@
     "file-loader": "^2.0.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.9.3",
-    "protractor": "~5.4.0",
     "sass-loader": "^7.1.0",
     "to-string-loader": "^1.1.5",
     "ts-node": "~7.0.0",
