+++ src/app/components/auth-form/auth-form.component.css
@@ -13,7 +13,7 @@
     box-shadow: 0px 2px 1px 1.5px rgb(200, 200, 200);
   }
     
-  .authorization-container > * {  
+  :host /deep/ .authorization-container > * {  
     width: 100%;
   }
   
@@ -28,3 +28,6 @@
     left: 60%;
     top: 85%; 
   }
+
+  
+
