+++ src/app/components/auth-form/auth-form.component.html
@@ -6,7 +6,7 @@
     </mat-form-field>
 
     <mat-form-field>
-        <input matInput placeholder="Password">
+        <input matInput placeholder="Password*">
     </mat-form-field>
 
     <ng-content>
@@ -14,8 +14,8 @@
     </ng-content>
     
     <div class="authorization_sign-container">
-        <button mat-button>SIGN UP</button>
-        <button mat-button>SIGN IN</button>
+        <button class="sign-container_up" mat-button>SIGN UP</button>
+        <button class="sign-container_in" mat-button>SIGN IN</button>
     </div>
     
 </div>
