+++ src/app/components/auth-form/auth-form.component.ts
@@ -1,0 +1,10 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-auth-form',
+  templateUrl: './auth-form.component.html',
+  styleUrls: ['./auth-form.component.css']
+})
+export class AuthFormComponent {
+  @Input() title: string;
+}
