+++ src/app/components/authentification.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-authentification',
+  templateUrl: './authentification.component.html'
+})
+export class AuthentificationComponent {
+
+}
