--- src/app/components/authentification/authentification.component.ts
@@ -1,10 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-authentification',
-  templateUrl: './authentification.component.html',
-  styleUrls: ['./authentification.component.css']
-})
-export class AuthentificationComponent {
-  @Input() title: string;
-}
