+++ src/app/components/authorization/authorization.component.css
@@ -1,0 +1,7 @@
+  :host /deep/ .sign-container_up{
+    order: 1;
+  }
+
+  :host /deep/ .sign-container_in{
+    order: 2;
+  }
