+++ src/app/components/authorization/authorization.component.html
@@ -1,1 +1,1 @@
-<app-authentification [title]="'LOGIN'"></app-authentification>
+<app-auth-form [title]="'LOGIN'"></app-auth-form>
