+++ src/app/components/registration/registration.component.html
@@ -1,5 +1,5 @@
-<app-authentification [title]="'SIGN UP'">
+<app-auth-form [title]="'SIGN UP'">
     <mat-form-field>
         <input matInput placeholder="Phone*">
     </mat-form-field>
-</app-authentification>
+</app-auth-form>
